requirements/dev.txt
-r common.txt
pytest-django==3.1.2
pytest-cov==2.5.1
coveralls==1.1
